remove solutions.ts empty file, move makeError to file
