+++ apps/nextjs-website/src/lib/makeError.ts
@@ -1,0 +1,3 @@
+export const makeError = ({ status, statusText }: Response) => {
+  return new Error(`${status} - ${statusText}`);
+};
